Pull up common members into abstract parent class
