+++ src/main/java/mil/dds/anet/beans/search/OrganizationSearchQuery.java
@@ -2,9 +2,8 @@ package mil.dds.anet.beans.search;
 
 import mil.dds.anet.beans.Organization.OrganizationType;
 
-public class OrganizationSearchQuery implements ISearchQuery {
+public class OrganizationSearchQuery extends AbstractSearchQuery {
 
-	String text;
 	OrganizationType type;
 	
 	//Search for organizations with a specific parent Org. 
@@ -13,22 +12,6 @@ public class OrganizationSearchQuery implements ISearchQuery {
 	//If true will include all orgs in the tree of the parentOrg
 	// Including the parent Org. 
 	Boolean parentOrgRecursively;
-	
-	int pageNum;
-	int pageSize;
-	
-	public OrganizationSearchQuery() { 
-		this.pageNum = 0;
-		this.pageSize = 10;
-	}
-	
-	public String getText() {
-		return text;
-	}
-	
-	public void setText(String text) {
-		this.text = text;
-	}
 	
 	public OrganizationType getType() {
 		return type;
@@ -54,27 +37,6 @@ public class OrganizationSearchQuery implements ISearchQuery {
 		this.parentOrgRecursively = parentOrgRecursively;
 	}
 
-	@Override
-	public int getPageNum() {
-		return pageNum;
-	}
-	
-	@Override
-	public void setPageNum(int pageNum) {
-		this.pageNum = pageNum;
-	}
-	
-	@Override
-	public int getPageSize() {
-		return pageSize;
-	}
-	
-	@Override
-	public void setPageSize(int pageSize) {
-		if (pageSize == 0) { return; } // that makes no sense. 
-		this.pageSize = pageSize;
-	}
-	
 	public static OrganizationSearchQuery withText(String text, int pageNum, int pageSize) {
 		OrganizationSearchQuery query = new OrganizationSearchQuery();
 		query.setText(text);
@@ -82,5 +44,5 @@ public class OrganizationSearchQuery implements ISearchQuery {
 		query.setPageSize(pageSize);
 		return query;
 	}
-	
+
 }
